Added database record keeping code
Added a 'DB' class with methods to query/update the database.
At this point, the text of the debuginfo and stacktrace is
stored directly in a database table, but attaching them to
a GitHub issue is going to require changes.  The table should be:
    create table upload_log (
	toolname char(16),
	version char(16),
	checksum char(128),		-- proper length???
	date timestamp,			-- proper data type???
	debuginfo varchar(1024),	-- proper length???
	stacktrace varchar(1024)	-- proper length???
    ) Type=InnoDB;
Put a primary key on <toolname,version,checksum>.